Address review comments on PathLayer data format
